Migrate Permission methods to ActiveRecord model
