+++ app/models/carto/organization_permission.rb
@@ -1,0 +1,17 @@
+module Carto
+  class OrganizationPermission
+
+    def add_read_permission(table)
+      table.add_organization_read_permission
+    end
+
+    def add_read_write_permission(table)
+      table.add_organization_read_write_permission
+    end
+
+    def is_permitted(table, access)
+      table.permission.permission_for_org == access
+    end
+
+  end
+end
